Move tests of resizing grouping column to grouping-column-test
